FastApi/requirements.txt
fastapi
uvicorn[standard]
httpx
python-keycloak
requests